Move extension code to app folder
